requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.1.0
aubio==0.4.9
bcrypt==4.2.1
beautifulsoup4==4.13.3
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
cryptography==44.0.0
deep-translator==1.11.4
defusedxml==0.7.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fastapi==0.115.8
flake8==7.1.1
flake8-black==0.3.6
flake8-bugbear==24.12.12
googletrans==4.0.2
gTTS==2.5.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jira==3.8.0
kombu==5.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
speedtest-cli==2.1.3
sqlparse==0.5.3
starlette==0.45.3
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
